refactor: create package for value objects
